Move Route and Router into their own files and export from 'index.js'
