+++ src/js/index.js
@@ -1,0 +1,7 @@
+import Router from './Router';
+
+export default function route(name, params, absolute, config) {
+    const router = new Router(name, params, absolute, config);
+
+    return name ? router.toString() : router;
+}
